+++ src/main/java/com/codeloon/ems/configuration/authentication/CustomerUserDetailsService.java
@@ -1,13 +1,11 @@
-package com.codeloon.ems.configuration;
+package com.codeloon.ems.configuration.authentication;
 
 
 import com.codeloon.ems.entity.User;
 import com.codeloon.ems.repository.UserRepository;
 import com.codeloon.ems.service.CustomUserDetails;
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
