+++ src/main/java/com/codeloon/ems/util/DataVarList.java
@@ -1,13 +1,17 @@
 package com.codeloon.ems.util;
 
-import org.springframework.stereotype.Component;
 
 public class DataVarList {
+    private DataVarList() {
+        throw new UnsupportedOperationException("DataVarList class cannot be instantiated");
+    }
+
+    // User roles
     public static final String ROLE_ADMIN = "ADMIN";
     public static final String ROLE_EMPLOYEE = "EMPLOYEE";
     public static final String ROLE_CLIENT = "CLIENT";
 
-
+    // Authentication responses,
     public static final String SUCCESS_AUTH="Authentication success.";
     public static final String AUTH_SUCCESS ="0";
     public static final String AUTH_FAILED_RESET_PASSWORD ="1";
