+++ src/main/java/com/codeloon/ems/util/ResponseCode.java
@@ -1,7 +1,10 @@
 package com.codeloon.ems.util;
 
 public class ResponseCode {
-    public static final String RSP_SUCCESS                          = "00";
-    public static final String RSP_ERROR                            = "01";
+    private ResponseCode() {
+        throw new UnsupportedOperationException("ResponseCode class cannot be instantiated");
+    }
+    public static final String RSP_SUCCESS = "00";
+    public static final String RSP_ERROR   = "01";
 
 }
